Strip message comments
Would move the comments into the message paragraphs if
they were in the middle of or just below a paragraph,
or re-format the comments as paragraphs e.g. "# A\n# B"
to "# A # B". Either way, comments would become part
of the message.

In the future, ignoring comments instead of stripping
them could be useful for some use cases, such as
commit-msg hooks running after this one.